requirements.txt
fastapi
uvicorn
jinja2
pyvis==0.3.1
python-multipart
pillow
gunicorn